+++ metadata.php
@@ -23,7 +23,7 @@ $aModule = [
         'en' => '',
     ],
     'thumbnail'   => 'logo.png',
-    'version'     => '4.0.0',
+    'version'     => '4.1.0',
     'author'      => 'OXID eSales AG & digidesk - media solutions',
     'url'         => 'https://www.oxid-esales.com',
     'email'       => 'info@oxid-esales.com',